- Move contents of old ContentView to SettingsView - Recreate ContentView.swift - Add Tab bar (Home/Settings)
